Settings screen rework

+++ app/src/main/java/com/joesemper/dronesettings/ui/settings/state/Limits.kt
@@ -1,6 +1,6 @@
 package com.joesemper.dronesettings.ui.settings.state
 
 data class Limits(
-    val minValue: Float,
-    val maxValue: Float
+    val minValue: Int,
+    val maxValue: Int
 )
